backend/requirements.txt
# Web framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.27
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
authlib==1.3.0
httpx==0.27.0
itsdangerous==2.1.2

# Garmin Connect
garminconnect==0.2.19
garth==0.4.14

# Encryption
cryptography==42.0.5

# Validation
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0

# Testing
pytest==8.0.1
pytest-asyncio==0.23.5
pytest-cov==4.1.0
pytest-httpx==0.28.0
factory-boy==3.3.0

# Utilities
python-dotenv==1.0.1
aiofiles==23.2.1
